Move ActivationCondition to it's own file.
This will be reused for table keyed map construction as well as glyph keyed.
