+++ ift/encoder/activation_condition.cc
@@ -1,0 +1,146 @@
+#include "ift/encoder/activation_condition.h"
+
+#include "common/int_set.h"
+
+using common::IntSet;
+using common::SegmentSet;
+
+namespace ift::encoder {
+
+ActivationCondition ActivationCondition::exclusive_segment(
+    segment_index_t index, patch_id_t activated) {
+  ActivationCondition condition;
+  condition.activated_ = activated;
+  condition.conditions_ = {{index}};
+  condition.is_exclusive_ = true;
+  return condition;
+}
+
+ActivationCondition ActivationCondition::and_segments(
+    const SegmentSet& segments, patch_id_t activated) {
+  ActivationCondition conditions;
+  conditions.activated_ = activated;
+
+  for (auto id : segments) {
+    conditions.conditions_.push_back(SegmentSet{id});
+  }
+
+  return conditions;
+}
+
+ActivationCondition ActivationCondition::or_segments(const SegmentSet& segments,
+                                                     patch_id_t activated,
+                                                     bool is_fallback) {
+  ActivationCondition conditions;
+  conditions.activated_ = activated;
+  conditions.conditions_.push_back(segments);
+  conditions.is_fallback_ = is_fallback;
+
+  return conditions;
+}
+
+ActivationCondition ActivationCondition::composite_condition(
+    absl::Span<const SegmentSet> groups, patch_id_t activated) {
+  ActivationCondition conditions;
+  conditions.activated_ = activated;
+  for (const auto& group : groups) {
+    conditions.conditions_.push_back(group);
+  }
+
+  return conditions;
+}
+
+std::string ActivationCondition::ToString() const {
+  std::stringstream out;
+  out << "if (";
+  bool first = true;
+  for (const auto& set : conditions()) {
+    if (!first) {
+      out << " AND ";
+    } else {
+      first = false;
+    }
+
+    if (set.size() > 1) {
+      out << "(";
+    }
+    bool first_inner = true;
+    for (uint32_t id : set) {
+      if (!first_inner) {
+        out << " OR ";
+      } else {
+        first_inner = false;
+      }
+      out << "s" << id;
+    }
+    if (set.size() > 1) {
+      out << ")";
+    }
+  }
+  out << ") then p" << activated();
+  return out.str();
+}
+
+bool ActivationCondition::operator<(const ActivationCondition& other) const {
+  if (conditions_.size() != other.conditions_.size()) {
+    return conditions_.size() < other.conditions_.size();
+  }
+
+  auto a = conditions_.begin();
+  auto b = other.conditions_.begin();
+  while (a != conditions_.end() && b != other.conditions_.end()) {
+    if (a->size() != b->size()) {
+      return a->size() < b->size();
+    }
+
+    auto aa = a->begin();
+    auto bb = b->begin();
+    while (aa != a->end() && bb != b->end()) {
+      if (*aa != *bb) {
+        return *aa < *bb;
+      }
+      aa++;
+      bb++;
+    }
+
+    a++;
+    b++;
+  }
+
+  if (activated_ != other.activated_) {
+    return activated_ < other.activated_;
+  }
+
+  if (is_exclusive_ != other.is_exclusive_) {
+    return is_exclusive_;
+  }
+
+  if (is_fallback_ != other.is_fallback_) {
+    return !is_fallback_;
+  }
+
+  // These two are equal
+  return false;
+}
+
+template <typename ProtoType>
+ProtoType ToSetProto(const IntSet& set) {
+  ProtoType values;
+  for (uint32_t v : set) {
+    values.add_values(v);
+  }
+  return values;
+}
+
+ActivationConditionProto ActivationCondition::ToConfigProto() const {
+  ActivationConditionProto proto;
+
+  for (const auto& ss : conditions()) {
+    *proto.add_required_segments() = ToSetProto<SegmentsProto>(ss);
+  }
+  proto.set_activated_patch(activated());
+
+  return proto;
+}
+
+}  // namespace ift::encoder
