+++ ift/encoder/types.h
@@ -1,0 +1,14 @@
+#ifndef IFT_ENCODER_TYPES_H_
+#define IFT_ENCODER_TYPES_H_
+
+#include <cstdint>
+
+namespace ift::encoder {
+
+typedef uint32_t segment_index_t;
+typedef uint32_t patch_id_t;
+typedef uint32_t glyph_id_t;
+
+}  // namespace ift::encoder
+
+#endif  // IFT_ENCODER_TYPES_H_
